requirements.txt.txt
streamlit
pillow
fpdf
requests